requirements.txt
absl-py==2.3.1
aif360==0.6.1
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
dash==3.1.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fairlearn==0.12.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.5
fpdf==1.7.2
gast==0.6.0
gevent==25.5.1
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.1
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
multipledispatch==1.0.0
murmurhash==1.0.13
namex==0.1.0
narwhals==1.46.0
natsort==8.4.0
nest-asyncio==1.6.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pandas_flavor==0.7.0
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
preshed==3.0.10
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyjanitor==0.31.0
pyparsing==3.2.3
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pyzmq==27.0.0
referencing==0.36.2
reportlab==4.4.2
requests==2.32.4
retrying==1.4.0
rich==14.0.0
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
streamlit==1.46.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsgi-oauth2==0.2.2
xarray==2025.7.1
zipp==3.23.0
zope.event==5.1
zope.interface==7.2
